Deploy to the polygon network and set the correct proxy addresses
It’s expensive to sell items on the ethereum network due to gas fees.
Instead, deploy to the polygon network for gas-free listings:

https://opensea.io/gas-free

According to this documentation, we need to set the proxyAddress to
OpenSea’s address:

https://docs.opensea.io/docs/polygon-basic-integration

The address for the main polygon network is given in that guide but I
had a hard time finding the correct one for the test network. A GitHub
search found the following, which has an address that seems to work on
the testnet.opensea.io:

https://gist.github.com/rahuldamodar94/6bdd022f3457934f2a104fd5f4bb45e4