Allow the contract owner to update puzzle and variant names
I want the contract to enforce that the rules of the game and its
probabilities are not allowed to change over time so that the rug can’t
be pulled from underneath people. However, I most likely will need to
add additional puzzles and variants over time as I develop puzzles in
the actual Worship the Sun game and introduce new artwork.