Add a ./bin/contract_size script
The contract size is close to exceeding 24KB which is the largest it can
be. Therefore, add a tool that allows me to see what effect different
code transformations have on the contract size so I can try to keep
below this limit.

In development, allow the contract size to exceed this amount so it
doesn’t halt progress and increase the runs (optimistically) to the
highest number it can be so that the code is optimised for maximum
performance and minimum gas costs rather than code size.