Change the action interfaces to take individual tokenIDs
Rather than put cards into slots inside the contract, I plan to remove
this functionality and instead call the actions with ids in a known
order to reduce the amount of work that goes on in the contract and make
it simpler so that it will take up less space.

For now, I'm shimming onto the existing code in each of these actions to
lessen the amount of code changing at once.