Move all ownership/double spend checks to the start
I want to extract an Abilities.sol library that performs the majority
of the checks for whether you can perform an action. This should reduce
the size of the contract (it is currently over the 24.5KB limit).

However, the balance checks need to remain inside this contract because
they access local storage which the library won’t have access to.