Add a script that generates an index of users’ decks
Unfortunately, there’s no out-of-the-box method to get the balance for
all tokens owned by a user. I initially tried adding a contract method
to loop through all card combinations but that turned out to be too
slow.

Instead, create a public/decks/ directory that contains an index of all
users’ decks of cards that they have purchased/combined/etc. I plan to
keep this index updated fairly often. In the browser, the user can then
fetch this deck as a starting point and I’ll then add a method to update
their deck for any new transactions added since the last checkpoint.