Request logs twice as quickly
We may as well try to fetch all batch/single logs for a block at once
because Promise.all preserves the order of the arrays when the promises
are resolved. This does mean we are more likely to hit Metamask’s RPC
rate limit (700/minute, 40/second - per IP) but I plan to build a
backoff mechanism to cope with that.

https://support.blockvigil.com/support/solutions/articles/67000346291-what-is-the-rate-limit-for-matic-rpc-