Ensure z-indexes are consistent for newly minted cards
When a card is minted, it should appear on top of the others that appear
over the outline. However, if the user is zoomed into a card at the time,
conceptually it is as though the user has picked up that card so the
newly minted card should not show over the top of it and when they zoom
out again, they should put the card back on top of the newly minted card.