Move already-visible cards over the card outline when minted
If a card stack is already in play and a new card is added to its stack
by performing an action, we should move it back over the card outline
so we can tell that it has changed. We can force React to treat it as a
new card by changing its key.