Make Filters return a shallow copy when something has changed
I want to use Filters in a setState, but I don’t want to leak loads of
memory by deep cloning the whole object every time something changes.
Therefore, when methods are called, change the same underlying state but
return a shallowCopy of the object so that React’s object identity check
fails and changes propagate.