Re-render cards after stacks are cleared from the hourglass area
I think these changes ensure no card stacks are skipped when paginating
if stacks are moved to/from the hourglass area and when the hourglass
area is cleared. This approach is conservative in that sometimes you’ll
see the last N cards on the current page become the first N cards on the
next page, but that’s better than unknowingly skipping past cards.