Rework how Master Copies are allocated to make it more inclusive
Firstly:

Previously, there were 10 Limited Editions and you had a 10% chance of
claiming one if all of the provided Star cards were Pristine. I don’t
like that this means you could go to the effort of hunting down 7
Pristine stars and not be rewarded for this effort. Instead, increase
the number of Limited Editions cards and always mint one if available.
If the user is savvy, they can check which puzzles are available before
trying to mint limited/master copy editions.

Secondly:

Instead of minting the Master Copy immediately to whoever gets there
first, make this a random lottery so that more users have a chance to
claim one. Otherwise, one user could easily grab all of them and exclude
everyone else. They could still do this if they write a script, but that
would likely become too expensive because of the large number of puzzles.
I want to encourage trading on OpenSea at Master tier to build a
community and this approach should lend itself better to that.