Make discard2Pickup1 apply the standard degrade behaviour
This makes the project more financially viable (£24k -> £35k/year)
without having to drastically increase the number of limited editions
(decreasing rarity) or change the condition probabilities. It’s slightly
weird in that if you pick up a new card, it probably makes more sense
for it to be eligible for Pristine condition again - but on the other
hand, this means the same degrade rule applies to every action which is
nice and consistent. Maybe this action should be renamed?