Try to make the filters easier to use…
- Don’t show counts of (0) next to options that aren’t relevant to the
current set of filters. Also, grey out these options slightly. These
options can still be selected but the filters are temporarily disabled
when this happens so a number of (0) is misleading. These counts are
wrong again if the user then selects an option that re-applies the
filter but I don’t think this is so bad because it’s usually more
obvious that a color filter was re-applied because they’re highlighted.

- Do a similar thing for variants. e.g. if ‘Active’ is selected then
grey out all types that don’t have an Active variant.

- Disable the color dropdown when door variants are selected.

- Disable the variant dropdown when color1 is selected at Master tier
because the only type with color at this tier is Star and it doesn’t
have variants.