Fix a bug with puzzles not being chosen uniformly random
Previously, we were randomly picking a series then picking a puzzle
inside that series, but there might be 100 puzzles in one series and 1
puzzle in another which would cause the latter puzzle to be chosen 100x
more often than the former puzzles.

Instead, use pickRandom to weight the selection based on how many
puzzles are in each series. When I made this change, one of the minting
tests started failing because it ran out of gas due to the extra
complexity of using pickRandom. Therefore, simplify pickRandom so that
it’s already given cumulative probability distributions to reduce the
amount of work back down again.

Also, the contract size exceeded its limit but I (fortunately) noticed
that two of the arguments to updateConstants aren’t used which brought
it back down again. So many constraints!