Re-arrange the DOM so that a corner can be cut without affecting video quality
If a video is placed inside a ancestor div that has a clip-path applied,
the video is rendered at a lower quality when a scale transform is
applied, i.e. when a card is picked up from the card table.

Therefore, move the puzzle video and the .types div that might contain
a video outside of the element that has the clip-path applied.