Reduce the file size by ~60% while losing almost no quality
I’m hesitant to apply any compression at all to images as these will be
displayed on OpenSea but a 95% quality level JPEG is virtually
indistinguishable from the lossless file unless you zoom in.

The file sizes go down from ~100KB -> 40KB which means that for 100
puzzles at ~8700 cards per puzzle, the total space needed is ~33GB which
is still a lot (too much to store on GitHub) but more manageable.