Make ./bin/generate_images concurrent
3 threads seems to be a good balance between speed and not thrashing the
system. 4 threads is ~8% faster but causes the system to lock a bit and
sometimes the browser instances crash and have to restart.

1000 screenshots took 17:14 with 3 threads, so about 1 per second. There
are ~8,700 cards per puzzle which would take about 2 and a half hours.
I’m hoping this will speed up when the images/videos are optimised.
Currently they’re raw images/videos which are really big.