Make it so cards can be minted in Reasonable condition
This makes it slightly harder to obtain master copies but I think it
adds more variety to the game as the player will get to see more of the
whimsical defects. It will also (hopefully) raise additional funds
which compensates for lowering the number of limited editions from 151
to 100 a while ago.