Set up a GitHub action to run the ./bin/update_index script
Creates a .github/workflows directory in the .gh-pages directory after
build. This adds a workflow that runs the indexer script then commits
the changes to the repository. The task is scheduled to run every 3
hours. The client side implementation can fetch any changes since then.

We should also run the indexer before deploying the website so that
there isn't a period of time where no index exists. This index should
probably be committed to the project repo so that it doesn't have to
rebuild from scratch every time which could take a long time.